# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Access Monitor API",
    "version": "1.0.0",
    "description": "JWT login for dashboard reads; per-device API key for Pi writes. All timestamps are server-generated. Import this file into Swagger Editor or Postman. Cookies are also accepted for dashboard reads."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "deviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "smart_access_token"
      }
    }
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "summary": "Current readings (aliases of the same handler)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "temperature": 24.6,
                  "humidity": 58,
                  "gas_status": "safe",
                  "fan_status": "off",
                  "device_online": true,
                  "last_seen": "2026-10-03T06:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown device",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Device not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          }
        ],
        "description": "Online means a valid reading was received within 2 minutes. Before first telemetry, readings/status strings and last_seen are null, online is false."
      }
    },
    "/api/dashboard/current": {
      "get": {
        "summary": "Current readings (aliases of the same handler)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "temperature": 24.6,
                  "humidity": 58,
                  "gas_status": "safe",
                  "fan_status": "off",
                  "device_online": true,
                  "last_seen": "2026-10-03T06:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown device",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Device not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          }
        ],
        "description": "Online means a valid reading was received within 2 minutes. Before first telemetry, readings/status strings and last_seen are null, online is false."
      }
    },
...
    "/api/readings": {
      "get": {
        "summary": "Latest N temperature/humidity readings, returned oldest first",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": [
                  {
                    "temperature": 24.6,
                    "humidity": 58,
                    "recorded_at": "2026-10-03T06:00:00Z",
                    "device_uid": "main-entrance-pi"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 50
            },
            "example": 50
          }
        ]
      }
    },
...
    "/api/alerts": {
      "get": {
        "summary": "Recent alerts, newest first",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "type": "gas_presence",
                    "message": "Gas detected",
                    "created_at": "2026-10-03T06:00:00Z",
                    "resolved_at": null,
                    "status": "Active",
                    "device_uid": "main-entrance-pi",
                    "device_name": "Main entrance Raspberry Pi"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 50
            },
            "example": 50
          }
        ],
        "description": "Types: high_temperature, high_humidity, gas_presence. Status is Active when resolved_at is null, otherwise Cleared."
      }
    },
    "/api/snapshots": {
      "post": {
        "summary": "Upload a camera image",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "file_path": "/uploads/example.png"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          },
          "413": {
            "description": "Too large",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Image must be at most 5 MB"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "device_uid",
                  "image"
                ],
                "properties": {
                  "device_uid": {
                    "type": "string",
                    "example": "main-entrance-pi"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "denied_access"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG or WebP, max 5 MB"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshots/latest": {
      "get": {
        "summary": "Latest snapshot metadata",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "file_path": "/uploads/example.png",
                  "reason": "denied_access",
                  "captured_at": "2026-10-03T06:00:00Z",
                  "device_name": "Main entrance Raspberry Pi"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          },
          "404": {
            "description": "No snapshot yet",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "No snapshots found"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          }
        ]
      }
    },
    "/uploads/{filename}": {
      "get": {
        "summary": "Read a snapshot image (login required)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example.png"
          }
        ],
        "responses": {
          "200": {
            "description": "Binary JPEG, PNG or WebP image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Login required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "404": {
            "description": "Image not found",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Endpoint not found"
                }
              }
            }
          }
        }
      }
    }
  }
}